backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
openai
pydub
moviepy<2.0
python-dotenv
tenacity
openai-whisper
google-generativeai
